RUM-4320 Shift the masking strategy for standard mappers
