+++ features/dd-sdk-android-session-replay/src/main/kotlin/com/datadog/android/sessionreplay/internal/recorder/mapper/CheckableTextViewMapper.kt
@@ -96,6 +96,10 @@ internal abstract class CheckableTextViewMapper<T>(
         )
     }
 
+    override fun resolveMaskedCheckable(view: T, mappingContext: MappingContext): List<MobileSegment.Wireframe>? {
+        return resolveNotCheckedCheckable(view, mappingContext)
+    }
+
     // endregion
 
     // region CheckableTextViewMapper
