--- features/dd-sdk-android-session-replay/src/main/kotlin/com/datadog/android/sessionreplay/internal/recorder/mapper/MaskCheckBoxMapper.kt
@@ -1,36 +1,0 @@
-/*
- * Unless explicitly stated otherwise all files in this repository are licensed under the Apache License Version 2.0.
- * This product includes software developed at Datadog (https://www.datadoghq.com/).
- * Copyright 2016-Present Datadog, Inc.
- */
-
-package com.datadog.android.sessionreplay.internal.recorder.mapper
-
-import android.widget.CheckBox
-import com.datadog.android.sessionreplay.model.MobileSegment
-import com.datadog.android.sessionreplay.utils.ColorStringFormatter
-import com.datadog.android.sessionreplay.utils.DrawableToColorMapper
-import com.datadog.android.sessionreplay.utils.ViewBoundsResolver
-import com.datadog.android.sessionreplay.utils.ViewIdentifierResolver
-
-internal class MaskCheckBoxMapper(
-    textWireframeMapper: TextViewMapper,
-    viewIdentifierResolver: ViewIdentifierResolver,
-    colorStringFormatter: ColorStringFormatter,
-    viewBoundsResolver: ViewBoundsResolver,
-    drawableToColorMapper: DrawableToColorMapper
-) : CheckBoxMapper(
-    textWireframeMapper,
-    viewIdentifierResolver,
-    colorStringFormatter,
-    viewBoundsResolver,
-    drawableToColorMapper
-) {
-
-    override fun resolveCheckedShapeStyle(view: CheckBox, checkBoxColor: String): MobileSegment.ShapeStyle? {
-        // in case the MASK rule is applied we do not want to show the selection in the
-        // checkbox related wireframe and in order to achieve that we need to provide
-        // a null value for the `ShapeStyle` and only keep the `ShapeBorder`.
-        return null
-    }
-}
