--- features/dd-sdk-android-session-replay/src/test/kotlin/com/datadog/android/sessionreplay/internal/recorder/mapper/MaskTextViewMapperTest.kt
@@ -1,56 +1,0 @@
-/*
- * Unless explicitly stated otherwise all files in this repository are licensed under the Apache License Version 2.0.
- * This product includes software developed at Datadog (https://www.datadoghq.com/).
- * Copyright 2016-Present Datadog, Inc.
- */
-
-package com.datadog.android.sessionreplay.internal.recorder.mapper
-
-import com.datadog.android.sessionreplay.forge.ForgeConfigurator
-import com.datadog.android.sessionreplay.internal.recorder.obfuscator.rules.MaskObfuscationRule
-import com.datadog.tools.unit.extensions.ApiLevelExtension
-import fr.xgouchet.elmyr.junit5.ForgeConfiguration
-import fr.xgouchet.elmyr.junit5.ForgeExtension
-import org.assertj.core.api.Assertions.assertThat
-import org.junit.jupiter.api.Test
-import org.junit.jupiter.api.extension.ExtendWith
-import org.junit.jupiter.api.extension.Extensions
-import org.mockito.junit.jupiter.MockitoExtension
-import org.mockito.junit.jupiter.MockitoSettings
-import org.mockito.quality.Strictness
-
-@Extensions(
-    ExtendWith(MockitoExtension::class),
-    ExtendWith(ForgeExtension::class),
-    ExtendWith(ApiLevelExtension::class)
-)
-@MockitoSettings(strictness = Strictness.LENIENT)
-@ForgeConfiguration(ForgeConfigurator::class)
-internal class MaskTextViewMapperTest : BaseTextViewWireframeMapperTest() {
-
-    override fun initTestedMapper(): TextViewMapper {
-        return MaskTextViewMapper(
-            mockViewIdentifierResolver,
-            mockColorStringFormatter,
-            mockViewBoundsResolver,
-            mockDrawableToColorMapper
-        ).apply {
-            textValueObfuscationRule = mockObfuscationRule
-        }
-    }
-
-    @Test
-    fun `M use the MaskObfuscationRule as defaultObfuscator when initialized`() {
-        // When
-        val textViewMapper = MaskTextViewMapper(
-            mockViewIdentifierResolver,
-            mockColorStringFormatter,
-            mockViewBoundsResolver,
-            mockDrawableToColorMapper
-        )
-
-        // Then
-        assertThat(textViewMapper.textValueObfuscationRule)
-            .isInstanceOf(MaskObfuscationRule::class.java)
-    }
-}
