+++ features/dd-sdk-android-session-replay/src/test/kotlin/com/datadog/android/sessionreplay/internal/recorder/mapper/NumberPickerMapperTest.kt
@@ -6,6 +6,7 @@
 
 package com.datadog.android.sessionreplay.internal.recorder.mapper
 
+import com.datadog.android.sessionreplay.SessionReplayPrivacy
 import com.datadog.android.sessionreplay.forge.ForgeConfigurator
 import fr.xgouchet.elmyr.Forge
 import fr.xgouchet.elmyr.junit5.ForgeConfiguration
@@ -36,9 +37,12 @@ internal class NumberPickerMapperTest : BaseNumberPickerMapperTest() {
         )
     }
 
+    // region privacy = ALLOW
+
     @Test
-    fun `M map the NumberPicker to a list of wireframes W map()`() {
+    fun `M return a list of wireframes W map() {privacy=ALLOW}`() {
         // Given
+        fakeMappingContext = fakeMappingContext.copy(privacy = SessionReplayPrivacy.ALLOW)
         val expectedSelectedLabelValue = fakeValue.toString()
         val expectedPrevLabelValue = (fakeValue - 1).toString()
         val expectedNextLabelValue = (fakeValue + 1).toString()
@@ -67,8 +71,9 @@ internal class NumberPickerMapperTest : BaseNumberPickerMapperTest() {
     }
 
     @Test
-    fun `M map the NumberPicker to a list of wireframes W map(){ value equals max }`() {
+    fun `M return a list of wireframes W map() {privacy=ALLOW, value=max }`() {
         // Given
+        fakeMappingContext = fakeMappingContext.copy(privacy = SessionReplayPrivacy.ALLOW)
         fakeValue = fakeMaxValue
         whenever(mockNumberPicker.value).thenReturn(fakeValue)
         val expectedSelectedLabelValue = fakeValue.toString()
@@ -99,8 +104,9 @@ internal class NumberPickerMapperTest : BaseNumberPickerMapperTest() {
     }
 
     @Test
-    fun `M map the NumberPicker to a list of wireframes W map(){ value equals min }`() {
+    fun `M return a list of wireframes W map() {privacy=ALLOW, value=min }`() {
         // Given
+        fakeMappingContext = fakeMappingContext.copy(privacy = SessionReplayPrivacy.ALLOW)
         fakeValue = fakeMinValue
         whenever(mockNumberPicker.value).thenReturn(fakeValue)
         val expectedSelectedLabelValue = fakeValue.toString()
@@ -131,10 +137,11 @@ internal class NumberPickerMapperTest : BaseNumberPickerMapperTest() {
     }
 
     @Test
-    fun `M map the NumberPicker to a list of wireframes W map() { displayedValues}`(
+    fun `M return a list of wireframes W map() {privacy=ALLOW, with displayedValues}`(
         forge: Forge
     ) {
         // Given
+        fakeMappingContext = fakeMappingContext.copy(privacy = SessionReplayPrivacy.ALLOW)
         val fakeDisplayedValues = forge.aList(size = (fakeMaxValue - fakeMinValue + 1)) { aString() }
             .toTypedArray()
         whenever(mockNumberPicker.displayedValues).thenReturn(fakeDisplayedValues)
@@ -167,10 +174,11 @@ internal class NumberPickerMapperTest : BaseNumberPickerMapperTest() {
     }
 
     @Test
-    fun `M map the NumberPicker to a list of wireframes W map(){ displayedValues, max value }`(
+    fun `M return a list of wireframes W map() {privacy=ALLOW, with displayedValues, value=max}`(
         forge: Forge
     ) {
         // Given
+        fakeMappingContext = fakeMappingContext.copy(privacy = SessionReplayPrivacy.ALLOW)
         fakeValue = fakeMaxValue
         whenever(mockNumberPicker.value).thenReturn(fakeValue)
         val fakeDisplayedValues = forge.aList(size = (fakeMaxValue - fakeMinValue + 1)) { aString() }
@@ -204,10 +212,11 @@ internal class NumberPickerMapperTest : BaseNumberPickerMapperTest() {
     }
 
     @Test
-    fun `M map the NumberPicker to a list of wireframes W map(){ displayedValues, min value }`(
+    fun `M return a list of wireframes W map() {privacy=ALLOW, with displayedValues, value=min}`(
         forge: Forge
     ) {
         // Given
+        fakeMappingContext = fakeMappingContext.copy(privacy = SessionReplayPrivacy.ALLOW)
         fakeValue = fakeMinValue
         whenever(mockNumberPicker.value).thenReturn(fakeValue)
         val fakeDisplayedValues = forge.aList(size = (fakeMaxValue - fakeMinValue + 1)) { aString() }
@@ -242,8 +251,9 @@ internal class NumberPickerMapperTest : BaseNumberPickerMapperTest() {
     }
 
     @Test
-    fun `M return empty list W map { prevLabelId null }`() {
+    fun `M return empty list W map {privacy=ALLOW, prevLabelId=null}`() {
         // Given
+        fakeMappingContext = fakeMappingContext.copy(privacy = SessionReplayPrivacy.ALLOW)
         whenever(
             mockViewIdentifierResolver
                 .resolveChildUniqueIdentifier(
@@ -261,8 +271,9 @@ internal class NumberPickerMapperTest : BaseNumberPickerMapperTest() {
     }
 
     @Test
-    fun `M return empty list W map { nextLabelId null }`() {
+    fun `M return empty list W map {privacy=ALLOW, nextLabelId=null}`() {
         // Given
+        fakeMappingContext = fakeMappingContext.copy(privacy = SessionReplayPrivacy.ALLOW)
         whenever(
             mockViewIdentifierResolver
                 .resolveChildUniqueIdentifier(
@@ -278,4 +289,328 @@ internal class NumberPickerMapperTest : BaseNumberPickerMapperTest() {
         // Then
         assertThat(wireframes).isEmpty()
     }
+
+    // endregion
+
+    // region privacy = MASK
+
+    @Test
+    fun `M return a list of wireframes W map() {privacy=MASK}`() {
+        // Given
+        fakeMappingContext = fakeMappingContext.copy(privacy = SessionReplayPrivacy.MASK)
+        val expectedSelectedLabelValue = "xxx"
+        val expectedTopDividerWireframe = fakeTopDividerWireframe()
+        val expectedSelectedLabelWireframe = fakeSelectedLabelWireframe()
+            .copy(text = expectedSelectedLabelValue)
+        val expectedBottomDividerWireframe = fakeBottomDividerWireframe()
+        // When
+        val wireframes = testedNumberPickerMapper.map(mockNumberPicker, fakeMappingContext, mockAsyncJobStatusCallback)
+
+        // Then
+        assertThat(wireframes).isEqualTo(
+            listOf(
+                expectedTopDividerWireframe,
+                expectedSelectedLabelWireframe,
+                expectedBottomDividerWireframe
+            )
+        )
+    }
+
+    @Test
+    fun `M return a list of wireframes W map() {privacy=MASK, value=max}`() {
+        // Given
+        fakeMappingContext = fakeMappingContext.copy(privacy = SessionReplayPrivacy.MASK)
+        fakeValue = fakeMaxValue
+        whenever(mockNumberPicker.value).thenReturn(fakeValue)
+        val expectedSelectedLabelValue = "xxx"
+        val expectedTopDividerWireframe = fakeTopDividerWireframe()
+        val expectedSelectedLabelWireframe = fakeSelectedLabelWireframe()
+            .copy(text = expectedSelectedLabelValue)
+        val expectedBottomDividerWireframe = fakeBottomDividerWireframe()
+        // When
+        val wireframes = testedNumberPickerMapper.map(mockNumberPicker, fakeMappingContext, mockAsyncJobStatusCallback)
+
+        // Then
+        assertThat(wireframes).isEqualTo(
+            listOf(
+                expectedTopDividerWireframe,
+                expectedSelectedLabelWireframe,
+                expectedBottomDividerWireframe
+            )
+        )
+    }
+
+    @Test
+    fun `M return a list of wireframes W map() {privacy=MASK, value=min }`() {
+        // Given
+        fakeMappingContext = fakeMappingContext.copy(privacy = SessionReplayPrivacy.MASK)
+        fakeValue = fakeMinValue
+        whenever(mockNumberPicker.value).thenReturn(fakeValue)
+        val expectedSelectedLabelValue = "xxx"
+        val expectedTopDividerWireframe = fakeTopDividerWireframe()
+        val expectedSelectedLabelWireframe = fakeSelectedLabelWireframe()
+            .copy(text = expectedSelectedLabelValue)
+        val expectedBottomDividerWireframe = fakeBottomDividerWireframe()
+
+        // When
+        val wireframes = testedNumberPickerMapper.map(mockNumberPicker, fakeMappingContext, mockAsyncJobStatusCallback)
+
+        // Then
+        assertThat(wireframes).isEqualTo(
+            listOf(
+                expectedTopDividerWireframe,
+                expectedSelectedLabelWireframe,
+                expectedBottomDividerWireframe
+            )
+        )
+    }
+
+    @Test
+    fun `M return a list of wireframes W map() {privacy=MASK, with displayedValues}`(
+        forge: Forge
+    ) {
+        // Given
+        fakeMappingContext = fakeMappingContext.copy(privacy = SessionReplayPrivacy.MASK)
+        val fakeDisplayedValues = forge.aList(size = (fakeMaxValue - fakeMinValue + 1)) { aString() }
+            .toTypedArray()
+        whenever(mockNumberPicker.displayedValues).thenReturn(fakeDisplayedValues)
+        val expectedSelectedLabelValue = "xxx"
+        val expectedTopDividerWireframe = fakeTopDividerWireframe()
+        val expectedSelectedLabelWireframe = fakeSelectedLabelWireframe()
+            .copy(text = expectedSelectedLabelValue)
+        val expectedBottomDividerWireframe = fakeBottomDividerWireframe()
+        // When
+        val wireframes = testedNumberPickerMapper.map(mockNumberPicker, fakeMappingContext, mockAsyncJobStatusCallback)
+
+        // Then
+        assertThat(wireframes).isEqualTo(
+            listOf(
+                expectedTopDividerWireframe,
+                expectedSelectedLabelWireframe,
+                expectedBottomDividerWireframe
+            )
+        )
+    }
+
+    @Test
+    fun `M return a list of wireframes W map() {privacy=MASK, with displayedValues, value=max}`(
+        forge: Forge
+    ) {
+        // Given
+        fakeMappingContext = fakeMappingContext.copy(privacy = SessionReplayPrivacy.MASK)
+        fakeValue = fakeMaxValue
+        whenever(mockNumberPicker.value).thenReturn(fakeValue)
+        val fakeDisplayedValues = forge.aList(size = (fakeMaxValue - fakeMinValue + 1)) { aString() }
+            .toTypedArray()
+        whenever(mockNumberPicker.displayedValues).thenReturn(fakeDisplayedValues)
+        val expectedSelectedLabelValue = "xxx"
+        val expectedTopDividerWireframe = fakeTopDividerWireframe()
+        val expectedSelectedLabelWireframe = fakeSelectedLabelWireframe()
+            .copy(text = expectedSelectedLabelValue)
+        val expectedBottomDividerWireframe = fakeBottomDividerWireframe()
+        // When
+        val wireframes = testedNumberPickerMapper.map(mockNumberPicker, fakeMappingContext, mockAsyncJobStatusCallback)
+
+        // Then
+        assertThat(wireframes).isEqualTo(
+            listOf(
+                expectedTopDividerWireframe,
+                expectedSelectedLabelWireframe,
+                expectedBottomDividerWireframe
+            )
+        )
+    }
+
+    @Test
+    fun `M return a list of wireframes W map() {privacy=MASK, with displayedValues, value=min}`(
+        forge: Forge
+    ) {
+        // Given
+        fakeMappingContext = fakeMappingContext.copy(privacy = SessionReplayPrivacy.MASK)
+        fakeValue = fakeMinValue
+        whenever(mockNumberPicker.value).thenReturn(fakeValue)
+        val fakeDisplayedValues = forge.aList(size = (fakeMaxValue - fakeMinValue + 1)) { aString() }
+            .toTypedArray()
+        whenever(mockNumberPicker.displayedValues).thenReturn(fakeDisplayedValues)
+        val expectedSelectedLabelValue = "xxx"
+        val expectedTopDividerWireframe = fakeTopDividerWireframe()
+        val expectedSelectedLabelWireframe = fakeSelectedLabelWireframe()
+            .copy(text = expectedSelectedLabelValue)
+        val expectedBottomDividerWireframe = fakeBottomDividerWireframe()
+
+        // When
+        val wireframes = testedNumberPickerMapper.map(mockNumberPicker, fakeMappingContext, mockAsyncJobStatusCallback)
+
+        // Then
+        assertThat(wireframes).isEqualTo(
+            listOf(
+                expectedTopDividerWireframe,
+                expectedSelectedLabelWireframe,
+                expectedBottomDividerWireframe
+            )
+        )
+    }
+
+    // endregion
+
+    // region privacy = MASK_USER_INPUT
+
+    @Test
+    fun `M return a list of wireframes W map() {privacy=MASK_USER_INPUT}`() {
+        // Given
+        fakeMappingContext = fakeMappingContext.copy(privacy = SessionReplayPrivacy.MASK_USER_INPUT)
+        val expectedSelectedLabelValue = "xxx"
+        val expectedTopDividerWireframe = fakeTopDividerWireframe()
+        val expectedSelectedLabelWireframe = fakeSelectedLabelWireframe()
+            .copy(text = expectedSelectedLabelValue)
+        val expectedBottomDividerWireframe = fakeBottomDividerWireframe()
+        // When
+        val wireframes = testedNumberPickerMapper.map(mockNumberPicker, fakeMappingContext, mockAsyncJobStatusCallback)
+
+        // Then
+        assertThat(wireframes).isEqualTo(
+            listOf(
+                expectedTopDividerWireframe,
+                expectedSelectedLabelWireframe,
+                expectedBottomDividerWireframe
+            )
+        )
+    }
+
+    @Test
+    fun `M return a list of wireframes W map() {privacy=MASK_USER_INPUT, value=max}`() {
+        // Given
+        fakeMappingContext = fakeMappingContext.copy(privacy = SessionReplayPrivacy.MASK_USER_INPUT)
+        fakeValue = fakeMaxValue
+        whenever(mockNumberPicker.value).thenReturn(fakeValue)
+        val expectedSelectedLabelValue = "xxx"
+        val expectedTopDividerWireframe = fakeTopDividerWireframe()
+        val expectedSelectedLabelWireframe = fakeSelectedLabelWireframe()
+            .copy(text = expectedSelectedLabelValue)
+        val expectedBottomDividerWireframe = fakeBottomDividerWireframe()
+        // When
+        val wireframes = testedNumberPickerMapper.map(mockNumberPicker, fakeMappingContext, mockAsyncJobStatusCallback)
+
+        // Then
+        assertThat(wireframes).isEqualTo(
+            listOf(
+                expectedTopDividerWireframe,
+                expectedSelectedLabelWireframe,
+                expectedBottomDividerWireframe
+            )
+        )
+    }
+
+    @Test
+    fun `M return a list of wireframes W map() {privacy=MASK_USER_INPUT, value=min }`() {
+        // Given
+        fakeMappingContext = fakeMappingContext.copy(privacy = SessionReplayPrivacy.MASK_USER_INPUT)
+        fakeValue = fakeMinValue
+        whenever(mockNumberPicker.value).thenReturn(fakeValue)
+        val expectedSelectedLabelValue = "xxx"
+        val expectedTopDividerWireframe = fakeTopDividerWireframe()
+        val expectedSelectedLabelWireframe = fakeSelectedLabelWireframe()
+            .copy(text = expectedSelectedLabelValue)
+        val expectedBottomDividerWireframe = fakeBottomDividerWireframe()
+
+        // When
+        val wireframes = testedNumberPickerMapper.map(mockNumberPicker, fakeMappingContext, mockAsyncJobStatusCallback)
+
+        // Then
+        assertThat(wireframes).isEqualTo(
+            listOf(
+                expectedTopDividerWireframe,
+                expectedSelectedLabelWireframe,
+                expectedBottomDividerWireframe
+            )
+        )
+    }
+
+    @Test
+    fun `M return a list of wireframes W map() {privacy=MASK_USER_INPUT, with displayedValues}`(
+        forge: Forge
+    ) {
+        // Given
+        fakeMappingContext = fakeMappingContext.copy(privacy = SessionReplayPrivacy.MASK_USER_INPUT)
+        val fakeDisplayedValues = forge.aList(size = (fakeMaxValue - fakeMinValue + 1)) { aString() }
+            .toTypedArray()
+        whenever(mockNumberPicker.displayedValues).thenReturn(fakeDisplayedValues)
+        val expectedSelectedLabelValue = "xxx"
+        val expectedTopDividerWireframe = fakeTopDividerWireframe()
+        val expectedSelectedLabelWireframe = fakeSelectedLabelWireframe()
+            .copy(text = expectedSelectedLabelValue)
+        val expectedBottomDividerWireframe = fakeBottomDividerWireframe()
+        // When
+        val wireframes = testedNumberPickerMapper.map(mockNumberPicker, fakeMappingContext, mockAsyncJobStatusCallback)
+
+        // Then
+        assertThat(wireframes).isEqualTo(
+            listOf(
+                expectedTopDividerWireframe,
+                expectedSelectedLabelWireframe,
+                expectedBottomDividerWireframe
+            )
+        )
+    }
+
+    @Test
+    fun `M return a list of wireframes W map() {privacy=MASK_USER_INPUT, with displayedValues, value=max}`(
+        forge: Forge
+    ) {
+        // Given
+        fakeMappingContext = fakeMappingContext.copy(privacy = SessionReplayPrivacy.MASK_USER_INPUT)
+        fakeValue = fakeMaxValue
+        whenever(mockNumberPicker.value).thenReturn(fakeValue)
+        val fakeDisplayedValues = forge.aList(size = (fakeMaxValue - fakeMinValue + 1)) { aString() }
+            .toTypedArray()
+        whenever(mockNumberPicker.displayedValues).thenReturn(fakeDisplayedValues)
+        val expectedSelectedLabelValue = "xxx"
+        val expectedTopDividerWireframe = fakeTopDividerWireframe()
+        val expectedSelectedLabelWireframe = fakeSelectedLabelWireframe()
+            .copy(text = expectedSelectedLabelValue)
+        val expectedBottomDividerWireframe = fakeBottomDividerWireframe()
+        // When
+        val wireframes = testedNumberPickerMapper.map(mockNumberPicker, fakeMappingContext, mockAsyncJobStatusCallback)
+
+        // Then
+        assertThat(wireframes).isEqualTo(
+            listOf(
+                expectedTopDividerWireframe,
+                expectedSelectedLabelWireframe,
+                expectedBottomDividerWireframe
+            )
+        )
+    }
+
+    @Test
+    fun `M return a list of wireframes W map() {privacy=MASK_USER_INPUT, with displayedValues, value=min}`(
+        forge: Forge
+    ) {
+        // Given
+        fakeMappingContext = fakeMappingContext.copy(privacy = SessionReplayPrivacy.MASK_USER_INPUT)
+        fakeValue = fakeMinValue
+        whenever(mockNumberPicker.value).thenReturn(fakeValue)
+        val fakeDisplayedValues = forge.aList(size = (fakeMaxValue - fakeMinValue + 1)) { aString() }
+            .toTypedArray()
+        whenever(mockNumberPicker.displayedValues).thenReturn(fakeDisplayedValues)
+        val expectedSelectedLabelValue = "xxx"
+        val expectedTopDividerWireframe = fakeTopDividerWireframe()
+        val expectedSelectedLabelWireframe = fakeSelectedLabelWireframe()
+            .copy(text = expectedSelectedLabelValue)
+        val expectedBottomDividerWireframe = fakeBottomDividerWireframe()
+
+        // When
+        val wireframes = testedNumberPickerMapper.map(mockNumberPicker, fakeMappingContext, mockAsyncJobStatusCallback)
+
+        // Then
+        assertThat(wireframes).isEqualTo(
+            listOf(
+                expectedTopDividerWireframe,
+                expectedSelectedLabelWireframe,
+                expectedBottomDividerWireframe
+            )
+        )
+    }
+
+    // endregion
 }
