+++ features/dd-sdk-android-session-replay/src/test/kotlin/com/datadog/android/sessionreplay/internal/recorder/mapper/SeekBarWireframeMapperTest.kt
@@ -9,6 +9,7 @@ package com.datadog.android.sessionreplay.internal.recorder.mapper
 import android.content.Context
 import android.content.res.Configuration
 import android.content.res.Resources
+import com.datadog.android.sessionreplay.SessionReplayPrivacy
 import com.datadog.android.sessionreplay.forge.ForgeConfigurator
 import com.datadog.android.sessionreplay.model.MobileSegment
 import fr.xgouchet.elmyr.Forge
@@ -41,9 +42,12 @@ internal class SeekBarWireframeMapperTest : BaseSeekBarWireframeMapperTest() {
         )
     }
 
+    // region privacy=ALLOW
+
     @Test
-    fun `M map the SeekBar to a list of wireframes W map()`() {
+    fun `M map the SeekBar to a list of wireframes W map() {privacy=ALLOW}`() {
         // Given
+        fakeMappingContext = fakeMappingContext.copy(privacy = SessionReplayPrivacy.ALLOW)
         val expectedInactiveTrackWireframe = MobileSegment.Wireframe.ShapeWireframe(
             id = fakeInactiveTrackId,
             x = fakeExpectedInactiveTrackXPos,
@@ -97,8 +101,9 @@ internal class SeekBarWireframeMapperTest : BaseSeekBarWireframeMapperTest() {
     }
 
     @Test
-    fun `M use default day color W map(progressTintList null, day theme)`(forge: Forge) {
+    fun `M use default day color W map() {privacy=ALLOW, progressTintList=null, day theme}`(forge: Forge) {
         // Given
+        fakeMappingContext = fakeMappingContext.copy(privacy = SessionReplayPrivacy.ALLOW)
         val fakeDefaultDayHtmlColor = forge.aStringMatching("#[0-9A-Fa-f]{8}")
         val fakeDefaultDayNotActiveHtmlColor = forge.aStringMatching("#[0-9A-Fa-f]{8}")
         whenever(
@@ -180,8 +185,9 @@ internal class SeekBarWireframeMapperTest : BaseSeekBarWireframeMapperTest() {
     }
 
     @Test
-    fun `M use default night color W map(progressTintList null, night theme)`(forge: Forge) {
+    fun `M use default night color W map() {privacy=ALLOW, progressTintList=null, night theme}`(forge: Forge) {
         // Given
+        fakeMappingContext = fakeMappingContext.copy(privacy = SessionReplayPrivacy.ALLOW)
         val fakeDefaultNightHtmlColor = forge.aStringMatching("#[0-9A-Fa-f]{8}")
         val fakeDefaultNightNotActiveHtmlColor = forge.aStringMatching("#[0-9A-Fa-f]{8}")
         whenever(
@@ -263,8 +269,9 @@ internal class SeekBarWireframeMapperTest : BaseSeekBarWireframeMapperTest() {
     }
 
     @Test
-    fun `M use default day color W map(thumbTintList null, day theme)`(forge: Forge) {
+    fun `M use default day color W map() {privacy=ALLOW, thumbTintList=null, day theme}`(forge: Forge) {
         // Given
+        fakeMappingContext = fakeMappingContext.copy(privacy = SessionReplayPrivacy.ALLOW)
         val fakeDefaultDayHtmlColor = forge.aStringMatching("#[0-9A-Fa-f]{8}")
         whenever(
             mockColorStringFormatter.formatColorAndAlphaAsHexString(
@@ -338,8 +345,9 @@ internal class SeekBarWireframeMapperTest : BaseSeekBarWireframeMapperTest() {
     }
 
     @Test
-    fun `M use default night color W map(thumbTintList null, night theme)`(forge: Forge) {
+    fun `M use default night color W map() {privacy=ALLOW, thumbTintList=null, night theme}`(forge: Forge) {
         // Given
+        fakeMappingContext = fakeMappingContext.copy(privacy = SessionReplayPrivacy.ALLOW)
         val fakeDefaultNightHtmlColor = forge.aStringMatching("#[0-9A-Fa-f]{8}")
         whenever(
             mockColorStringFormatter.formatColorAndAlphaAsHexString(
@@ -411,4 +419,270 @@ internal class SeekBarWireframeMapperTest : BaseSeekBarWireframeMapperTest() {
             )
         )
     }
+
+    // endregion
+
+    // region privacy=MASK
+
+    @Test
+    fun `M map the SeekBar to a list of wireframes W map() {privacy=MASK}`() {
+        // Given
+        fakeMappingContext = fakeMappingContext.copy(privacy = SessionReplayPrivacy.MASK)
+        val expectedInactiveTrackWireframe = MobileSegment.Wireframe.ShapeWireframe(
+            id = fakeInactiveTrackId,
+            x = fakeExpectedInactiveTrackXPos,
+            y = fakeExpectedInactiveTrackYPos,
+            width = fakeExpectedInactiveTrackWidth,
+            height = fakeExpectedInactiveTrackHeight,
+            shapeStyle = MobileSegment.ShapeStyle(
+                backgroundColor = fakeExpectedTrackInactiveHtmlColor,
+                opacity = fakeViewAlpha
+            )
+        )
+
+        // When
+        val mappedWireframes = testedSeekBarWireframeMapper.map(
+            mockSeekBar,
+            fakeMappingContext,
+            mockAsyncJobStatusCallback
+        )
+
+        // Then
+        assertThat(mappedWireframes).isEqualTo(
+            listOf(
+                expectedInactiveTrackWireframe
+            )
+        )
+    }
+
+    @Test
+    fun `M use default day color W map() {privacy=MASK, progressTintList=null, day theme}`(forge: Forge) {
+        // Given
+        fakeMappingContext = fakeMappingContext.copy(privacy = SessionReplayPrivacy.MASK)
+        val fakeDefaultDayNotActiveHtmlColor = forge.aStringMatching("#[0-9A-Fa-f]{8}")
+        whenever(
+            mockColorStringFormatter.formatColorAndAlphaAsHexString(
+                SeekBarWireframeMapper.DAY_MODE_COLOR,
+                SeekBarWireframeMapper.PARTIALLY_OPAQUE_ALPHA_VALUE
+            )
+        )
+            .thenReturn(fakeDefaultDayNotActiveHtmlColor)
+        whenever(mockSeekBar.progressTintList).thenReturn(null)
+        val mockResources = mock<Resources> {
+            whenever(it.configuration).thenReturn(
+                Configuration().apply {
+                    this.uiMode =
+                        Configuration.UI_MODE_NIGHT_NO
+                }
+            )
+        }
+        val mockContext = mock<Context> {
+            whenever(it.resources).thenReturn(mockResources)
+        }
+        whenever(mockSeekBar.context).thenReturn(mockContext)
+        val expectedInactiveTrackWireframe = MobileSegment.Wireframe.ShapeWireframe(
+            id = fakeInactiveTrackId,
+            x = fakeExpectedInactiveTrackXPos,
+            y = fakeExpectedInactiveTrackYPos,
+            width = fakeExpectedInactiveTrackWidth,
+            height = fakeExpectedInactiveTrackHeight,
+            shapeStyle = MobileSegment.ShapeStyle(
+                backgroundColor = fakeDefaultDayNotActiveHtmlColor,
+                opacity = fakeViewAlpha
+            )
+        )
+
+        // When
+        val mappedWireframes = testedSeekBarWireframeMapper.map(
+            mockSeekBar,
+            fakeMappingContext,
+            mockAsyncJobStatusCallback
+        )
+
+        // Then
+        assertThat(mappedWireframes).isEqualTo(listOf(expectedInactiveTrackWireframe))
+    }
+
+    @Test
+    fun `M use default night color W map() {privacy=MASK, progressTintList=null, night theme}`(forge: Forge) {
+        // Given
+        fakeMappingContext = fakeMappingContext.copy(privacy = SessionReplayPrivacy.MASK)
+        val fakeDefaultNightNotActiveHtmlColor = forge.aStringMatching("#[0-9A-Fa-f]{8}")
+        whenever(
+            mockColorStringFormatter.formatColorAndAlphaAsHexString(
+                SeekBarWireframeMapper.NIGHT_MODE_COLOR,
+                SeekBarWireframeMapper.PARTIALLY_OPAQUE_ALPHA_VALUE
+            )
+        )
+            .thenReturn(fakeDefaultNightNotActiveHtmlColor)
+        whenever(mockSeekBar.progressTintList).thenReturn(null)
+        val mockResources = mock<Resources> {
+            whenever(it.configuration).thenReturn(
+                Configuration().apply {
+                    this.uiMode =
+                        Configuration.UI_MODE_NIGHT_YES
+                }
+            )
+        }
+        val mockContext = mock<Context> {
+            whenever(it.resources).thenReturn(mockResources)
+        }
+        whenever(mockSeekBar.context).thenReturn(mockContext)
+        val expectedInactiveTrackWireframe = MobileSegment.Wireframe.ShapeWireframe(
+            id = fakeInactiveTrackId,
+            x = fakeExpectedInactiveTrackXPos,
+            y = fakeExpectedInactiveTrackYPos,
+            width = fakeExpectedInactiveTrackWidth,
+            height = fakeExpectedInactiveTrackHeight,
+            shapeStyle = MobileSegment.ShapeStyle(
+                backgroundColor = fakeDefaultNightNotActiveHtmlColor,
+                opacity = fakeViewAlpha
+            )
+        )
+
+        // When
+        val mappedWireframes = testedSeekBarWireframeMapper.map(
+            mockSeekBar,
+            fakeMappingContext,
+            mockAsyncJobStatusCallback
+        )
+
+        // Then
+        assertThat(mappedWireframes).isEqualTo(listOf(expectedInactiveTrackWireframe))
+    }
+
+    // endregion
+
+    // region privacy=MASK
+
+    @Test
+    fun `M map the SeekBar to a list of wireframes W map() {privacy=MASK_USER_INPUT}`() {
+        // Given
+        fakeMappingContext = fakeMappingContext.copy(privacy = SessionReplayPrivacy.MASK_USER_INPUT)
+        val expectedInactiveTrackWireframe = MobileSegment.Wireframe.ShapeWireframe(
+            id = fakeInactiveTrackId,
+            x = fakeExpectedInactiveTrackXPos,
+            y = fakeExpectedInactiveTrackYPos,
+            width = fakeExpectedInactiveTrackWidth,
+            height = fakeExpectedInactiveTrackHeight,
+            shapeStyle = MobileSegment.ShapeStyle(
+                backgroundColor = fakeExpectedTrackInactiveHtmlColor,
+                opacity = fakeViewAlpha
+            )
+        )
+
+        // When
+        val mappedWireframes = testedSeekBarWireframeMapper.map(
+            mockSeekBar,
+            fakeMappingContext,
+            mockAsyncJobStatusCallback
+        )
+
+        // Then
+        assertThat(mappedWireframes).isEqualTo(
+            listOf(
+                expectedInactiveTrackWireframe
+            )
+        )
+    }
+
+    @Test
+    fun `M use default day color W map() {privacy=MASK_USER_INPUT, progressTintList=null, day theme}`(forge: Forge) {
+        // Given
+        fakeMappingContext = fakeMappingContext.copy(privacy = SessionReplayPrivacy.MASK_USER_INPUT)
+        val fakeDefaultDayNotActiveHtmlColor = forge.aStringMatching("#[0-9A-Fa-f]{8}")
+        whenever(
+            mockColorStringFormatter.formatColorAndAlphaAsHexString(
+                SeekBarWireframeMapper.DAY_MODE_COLOR,
+                SeekBarWireframeMapper.PARTIALLY_OPAQUE_ALPHA_VALUE
+            )
+        )
+            .thenReturn(fakeDefaultDayNotActiveHtmlColor)
+        whenever(mockSeekBar.progressTintList).thenReturn(null)
+        val mockResources = mock<Resources> {
+            whenever(it.configuration).thenReturn(
+                Configuration().apply {
+                    this.uiMode =
+                        Configuration.UI_MODE_NIGHT_NO
+                }
+            )
+        }
+        val mockContext = mock<Context> {
+            whenever(it.resources).thenReturn(mockResources)
+        }
+        whenever(mockSeekBar.context).thenReturn(mockContext)
+        val expectedInactiveTrackWireframe = MobileSegment.Wireframe.ShapeWireframe(
+            id = fakeInactiveTrackId,
+            x = fakeExpectedInactiveTrackXPos,
+            y = fakeExpectedInactiveTrackYPos,
+            width = fakeExpectedInactiveTrackWidth,
+            height = fakeExpectedInactiveTrackHeight,
+            shapeStyle = MobileSegment.ShapeStyle(
+                backgroundColor = fakeDefaultDayNotActiveHtmlColor,
+                opacity = fakeViewAlpha
+            )
+        )
+
+        // When
+        val mappedWireframes = testedSeekBarWireframeMapper.map(
+            mockSeekBar,
+            fakeMappingContext,
+            mockAsyncJobStatusCallback
+        )
+
+        // Then
+        assertThat(mappedWireframes).isEqualTo(listOf(expectedInactiveTrackWireframe))
+    }
+
+    @Test
+    fun `M use default night color W map() {privacy=MASK_USER_INPUT, progressTintList=null, night theme}`(
+        forge: Forge
+    ) {
+        // Given
+        fakeMappingContext = fakeMappingContext.copy(privacy = SessionReplayPrivacy.MASK_USER_INPUT)
+        val fakeDefaultNightNotActiveHtmlColor = forge.aStringMatching("#[0-9A-Fa-f]{8}")
+        whenever(
+            mockColorStringFormatter.formatColorAndAlphaAsHexString(
+                SeekBarWireframeMapper.NIGHT_MODE_COLOR,
+                SeekBarWireframeMapper.PARTIALLY_OPAQUE_ALPHA_VALUE
+            )
+        )
+            .thenReturn(fakeDefaultNightNotActiveHtmlColor)
+        whenever(mockSeekBar.progressTintList).thenReturn(null)
+        val mockResources = mock<Resources> {
+            whenever(it.configuration).thenReturn(
+                Configuration().apply {
+                    this.uiMode =
+                        Configuration.UI_MODE_NIGHT_YES
+                }
+            )
+        }
+        val mockContext = mock<Context> {
+            whenever(it.resources).thenReturn(mockResources)
+        }
+        whenever(mockSeekBar.context).thenReturn(mockContext)
+        val expectedInactiveTrackWireframe = MobileSegment.Wireframe.ShapeWireframe(
+            id = fakeInactiveTrackId,
+            x = fakeExpectedInactiveTrackXPos,
+            y = fakeExpectedInactiveTrackYPos,
+            width = fakeExpectedInactiveTrackWidth,
+            height = fakeExpectedInactiveTrackHeight,
+            shapeStyle = MobileSegment.ShapeStyle(
+                backgroundColor = fakeDefaultNightNotActiveHtmlColor,
+                opacity = fakeViewAlpha
+            )
+        )
+
+        // When
+        val mappedWireframes = testedSeekBarWireframeMapper.map(
+            mockSeekBar,
+            fakeMappingContext,
+            mockAsyncJobStatusCallback
+        )
+
+        // Then
+        assertThat(mappedWireframes).isEqualTo(listOf(expectedInactiveTrackWireframe))
+    }
+
+    // endregion
 }
