+++ features/dd-sdk-android-session-replay/src/test/kotlin/com/datadog/android/sessionreplay/internal/recorder/mapper/SwitchCompatMapperTest.kt
@@ -6,6 +6,7 @@
 
 package com.datadog.android.sessionreplay.internal.recorder.mapper
 
+import com.datadog.android.sessionreplay.SessionReplayPrivacy
 import com.datadog.android.sessionreplay.forge.ForgeConfigurator
 import com.datadog.android.sessionreplay.model.MobileSegment
 import fr.xgouchet.elmyr.Forge
@@ -81,8 +82,13 @@ internal class SwitchCompatMapperTest : BaseSwitchCompatMapperTest() {
         )
 
         // Then
-        assertThat(resolvedWireframes)
-            .isEqualTo(fakeTextWireframes + expectedTrackWireframe + expectedThumbWireframe)
+        if (fakeMappingContext.privacy == SessionReplayPrivacy.ALLOW) {
+            assertThat(resolvedWireframes)
+                .isEqualTo(fakeTextWireframes + expectedTrackWireframe + expectedThumbWireframe)
+        } else {
+            assertThat(resolvedWireframes)
+                .isEqualTo(fakeTextWireframes + expectedTrackWireframe)
+        }
     }
 
     @Test
@@ -128,8 +134,13 @@ internal class SwitchCompatMapperTest : BaseSwitchCompatMapperTest() {
         )
 
         // Then
-        assertThat(resolvedWireframes)
-            .isEqualTo(fakeTextWireframes + expectedTrackWireframe + expectedThumbWireframe)
+        if (fakeMappingContext.privacy == SessionReplayPrivacy.ALLOW) {
+            assertThat(resolvedWireframes)
+                .isEqualTo(fakeTextWireframes + expectedTrackWireframe + expectedThumbWireframe)
+        } else {
+            assertThat(resolvedWireframes)
+                .isEqualTo(fakeTextWireframes + expectedTrackWireframe)
+        }
     }
 
     @Test
@@ -144,6 +155,21 @@ internal class SwitchCompatMapperTest : BaseSwitchCompatMapperTest() {
             )
         ).thenReturn(null)
         whenever(mockSwitch.isChecked).thenReturn(forge.aBool())
+        val expectedThumbWidth = normalizedThumbWidth - normalizedThumbRightPadding - normalizedThumbLeftPadding
+        val expectedTrackWidth = expectedThumbWidth * 2
+        val expectedTrackHeight = normalizedTrackHeight - normalizedThumbRightPadding - normalizedThumbLeftPadding
+        val expectedTrackWireframe = MobileSegment.Wireframe.ShapeWireframe(
+            id = fakeTrackIdentifier,
+            x = fakeViewGlobalBounds.x + fakeViewGlobalBounds.width - expectedTrackWidth,
+            y = fakeViewGlobalBounds.y + (fakeViewGlobalBounds.height - expectedTrackHeight) / 2,
+            width = expectedTrackWidth,
+            height = expectedTrackHeight,
+            border = null,
+            shapeStyle = MobileSegment.ShapeStyle(
+                backgroundColor = fakeCurrentTextColorString,
+                mockSwitch.alpha
+            )
+        )
 
         // When
         val resolvedWireframes = testedSwitchCompatMapper.map(
@@ -153,6 +179,7 @@ internal class SwitchCompatMapperTest : BaseSwitchCompatMapperTest() {
         )
 
         // Then
-        assertThat(resolvedWireframes).isEqualTo(fakeTextWireframes)
+        assertThat(resolvedWireframes)
+            .isEqualTo(fakeTextWireframes + expectedTrackWireframe)
     }
 }
