+++ features/dd-sdk-android-session-replay/src/test/kotlin/com/datadog/android/sessionreplay/internal/recorder/mapper/TextViewMapperTest.kt
@@ -6,17 +6,50 @@
 
 package com.datadog.android.sessionreplay.internal.recorder.mapper
 
+import android.content.res.ColorStateList
+import android.content.res.Resources
+import android.graphics.Typeface
+import android.graphics.drawable.ColorDrawable
+import android.graphics.drawable.Drawable
+import android.util.DisplayMetrics
+import android.widget.TextView
+import com.datadog.android.sessionreplay.SessionReplayPrivacy
 import com.datadog.android.sessionreplay.forge.ForgeConfigurator
+import com.datadog.android.sessionreplay.internal.recorder.aMockTextView
+import com.datadog.android.sessionreplay.internal.recorder.densityNormalized
 import com.datadog.android.sessionreplay.internal.recorder.obfuscator.rules.AllowObfuscationRule
+import com.datadog.android.sessionreplay.internal.recorder.obfuscator.rules.MaskInputObfuscationRule
+import com.datadog.android.sessionreplay.internal.recorder.obfuscator.rules.MaskObfuscationRule
+import com.datadog.android.sessionreplay.internal.utils.shapeStyle
+import com.datadog.android.sessionreplay.model.MobileSegment
+import com.datadog.android.sessionreplay.utils.GlobalBounds
+import com.datadog.android.sessionreplay.utils.ImageWireframeHelper
 import com.datadog.tools.unit.extensions.ApiLevelExtension
+import fr.xgouchet.elmyr.Forge
+import fr.xgouchet.elmyr.annotation.FloatForgery
+import fr.xgouchet.elmyr.annotation.Forgery
+import fr.xgouchet.elmyr.annotation.IntForgery
+import fr.xgouchet.elmyr.annotation.LongForgery
+import fr.xgouchet.elmyr.annotation.StringForgery
 import fr.xgouchet.elmyr.junit5.ForgeConfiguration
 import fr.xgouchet.elmyr.junit5.ForgeExtension
 import org.assertj.core.api.Assertions.assertThat
+import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
 import org.junit.jupiter.api.extension.Extensions
+import org.junit.jupiter.params.ParameterizedTest
+import org.junit.jupiter.params.provider.MethodSource
+import org.mockito.Mock
 import org.mockito.junit.jupiter.MockitoExtension
 import org.mockito.junit.jupiter.MockitoSettings
+import org.mockito.kotlin.any
+import org.mockito.kotlin.anyOrNull
+import org.mockito.kotlin.doReturn
+import org.mockito.kotlin.eq
+import org.mockito.kotlin.mock
+import org.mockito.kotlin.verify
+import org.mockito.kotlin.whenever
 import org.mockito.quality.Strictness
 
 @Extensions(
@@ -26,32 +59,516 @@ import org.mockito.quality.Strictness
 )
 @MockitoSettings(strictness = Strictness.LENIENT)
 @ForgeConfiguration(ForgeConfigurator::class)
-internal class TextViewMapperTest : BaseTextViewWireframeMapperTest() {
+internal class TextViewMapperTest : BaseWireframeMapperTest() {
 
-    override fun initTestedMapper(): TextViewMapper {
-        return TextViewMapper(
-            mockObfuscationRule,
-            mockViewIdentifierResolver,
-            mockColorStringFormatter,
-            mockViewBoundsResolver,
-            mockDrawableToColorMapper
-        ).apply {
-            textValueObfuscationRule = mockObfuscationRule
-        }
-    }
+    private lateinit var testedTextWireframeMapper: TextViewMapper
 
-    @Test
-    fun `M use the AllowObfuscationRule when initialized`() {
-        // When
-        val textViewMapper = TextViewMapper(
+    @Mock
+    lateinit var mockResources: Resources
+
+    @Mock
+    lateinit var mockImageWireframeHelper: ImageWireframeHelper
+
+    @Mock
+    lateinit var mockDisplayMetrics: DisplayMetrics
+
+    @StringForgery
+    lateinit var fakeText: String
+
+    @LongForgery
+    var fakeWireframeId: Long = 0
+
+    @IntForgery(min = 0, max = 0xffffff)
+    var fakeCurrentTextColor: Int = 0
+
+    @StringForgery(regex = "#[0-9A-F]{8}")
+    lateinit var fakeCurrentTextColorString: String
+
+    @Forgery
+    lateinit var fakeBounds: GlobalBounds
+
+    @BeforeEach
+    fun `set up`() {
+        fakeMappingContext = fakeMappingContext.copy(imageWireframeHelper = mockImageWireframeHelper)
+        whenever(mockViewIdentifierResolver.resolveChildUniqueIdentifier(any(), any()))
+            .thenReturn(System.identityHashCode(this).toLong())
+
+        whenever(mockResources.displayMetrics).thenReturn(mockDisplayMetrics)
+
+        whenever(mockViewIdentifierResolver.resolveViewId(any())) doReturn fakeWireframeId
+
+        whenever(mockViewBoundsResolver.resolveViewGlobalBounds(any(), any()))
+            .thenReturn(fakeBounds)
+
+        whenever(
+            mockColorStringFormatter.formatColorAndAlphaAsHexString(
+                fakeCurrentTextColor,
+                OPAQUE_ALPHA_VALUE
+            )
+        ) doReturn fakeCurrentTextColorString
+
+        testedTextWireframeMapper = TextViewMapper(
             mockViewIdentifierResolver,
             mockColorStringFormatter,
             mockViewBoundsResolver,
             mockDrawableToColorMapper
         )
+    }
+
+    private fun TextView.toTextWireframes(): List<MobileSegment.Wireframe.TextWireframe> {
+        val screenDensity = fakeMappingContext.systemInformation.screenDensity
+        return listOf(
+            MobileSegment.Wireframe.TextWireframe(
+                id = fakeWireframeId,
+                x = fakeBounds.x,
+                y = fakeBounds.y,
+                width = fakeBounds.width,
+                height = fakeBounds.height,
+                text = resolveTextValue(this),
+                textStyle = MobileSegment.TextStyle(
+                    TextViewMapper.SANS_SERIF_FAMILY_NAME,
+                    textSize.toLong().densityNormalized(screenDensity),
+                    fakeCurrentTextColorString
+                ),
+                textPosition = MobileSegment.TextPosition(
+                    MobileSegment.Padding(0, 0, 0, 0),
+                    alignment =
+                    MobileSegment.Alignment(
+                        MobileSegment.Horizontal.LEFT,
+                        MobileSegment.Vertical
+                            .CENTER
+                    )
+                )
+            )
+        )
+    }
+
+    @ParameterizedTest
+    @MethodSource("textTypefaces")
+    fun `M resolve a TextWireframe W map() { TextView with fontStyle }`(
+        fakeTypeface: Typeface?,
+        expectedFontFamily: String,
+        forge: Forge
+    ) {
+        // Given
+        val fakeFontSize = forge.aFloat(min = 0f)
+        val mockTextView: TextView = forge.aMockTextView().apply {
+            whenever(this.typeface).thenReturn(fakeTypeface)
+            whenever(this.textSize).thenReturn(fakeFontSize)
+            whenever(this.currentTextColor).thenReturn(fakeCurrentTextColor)
+            whenever(this.text).thenReturn(fakeText)
+            whenever(this.resources).thenReturn(mockResources)
+        }
+
+        // When
+        val textWireframes = testedTextWireframeMapper.map(mockTextView, fakeMappingContext, mockAsyncJobStatusCallback)
 
         // Then
-        assertThat(textViewMapper.textValueObfuscationRule)
-            .isInstanceOf(AllowObfuscationRule::class.java)
+        val expectedWireframes = mockTextView.toTextWireframes().map {
+            it.copy(
+                text = resolveExpectedText(mockTextView),
+                textStyle = MobileSegment.TextStyle(
+                    expectedFontFamily,
+                    fakeFontSize.toLong().densityNormalized(fakeMappingContext.systemInformation.screenDensity),
+                    fakeCurrentTextColorString
+                )
+            )
+        }
+        assertThat(textWireframes).isEqualTo(expectedWireframes)
+    }
+
+    @ParameterizedTest
+    @MethodSource("textAlignments")
+    fun `M resolve a TextWireframe W map() { TextView with textAlignment }`(
+        fakeTextAlignment: Int,
+        expectedTextAlignment: MobileSegment.Alignment,
+        forge: Forge
+    ) {
+        // Given
+        val mockTextView: TextView = forge.aMockTextView().apply {
+            whenever(this.text).thenReturn(fakeText)
+            whenever(this.typeface).thenReturn(mock())
+            whenever(this.textAlignment).thenReturn(fakeTextAlignment)
+            whenever(this.resources).thenReturn(mockResources)
+            whenever(this.currentTextColor).thenReturn(fakeCurrentTextColor)
+        }
+
+        // When
+        val textWireframes = testedTextWireframeMapper.map(mockTextView, fakeMappingContext, mockAsyncJobStatusCallback)
+
+        // Then
+        val expectedWireframes = mockTextView.toTextWireframes().map {
+            it.copy(
+                text = resolveExpectedText(mockTextView),
+                textPosition = MobileSegment.TextPosition(
+                    padding = MobileSegment.Padding(0, 0, 0, 0),
+                    alignment = expectedTextAlignment
+                )
+            )
+        }
+        assertThat(textWireframes).isEqualTo(expectedWireframes)
+    }
+
+    @ParameterizedTest
+    @MethodSource("textAlignmentsFromGravity")
+    fun `M resolve a TextWireframe W map() { TextView with textAlignment from gravity }`(
+        fakeGravity: Int,
+        expectedTextAlignment: MobileSegment.Alignment,
+        forge: Forge
+    ) {
+        // Given
+        val mockTextView: TextView = forge.aMockTextView().apply {
+            whenever(this.text).thenReturn(fakeText)
+            whenever(this.typeface).thenReturn(mock())
+            whenever(this.textAlignment).thenReturn(TextView.TEXT_ALIGNMENT_GRAVITY)
+            whenever(this.gravity).thenReturn(fakeGravity)
+            whenever(this.resources).thenReturn(mockResources)
+            whenever(this.currentTextColor).thenReturn(fakeCurrentTextColor)
+        }
+
+        // When
+        val textWireframes = testedTextWireframeMapper.map(mockTextView, fakeMappingContext, mockAsyncJobStatusCallback)
+
+        // Then
+        val expectedWireframes = mockTextView.toTextWireframes().map {
+            it.copy(
+                text = resolveExpectedText(mockTextView),
+                textPosition = MobileSegment.TextPosition(
+                    padding = MobileSegment.Padding(0, 0, 0, 0),
+                    alignment = expectedTextAlignment
+                )
+            )
+        }
+        assertThat(textWireframes).isEqualTo(expectedWireframes)
+    }
+
+    @Test
+    fun `M resolve a TextWireframe W map() { TextView with textPadding }`(
+        forge: Forge
+    ) {
+        // Given
+        val fakeTextPaddingTop = forge.anInt()
+        val fakeTextPaddingBottom = forge.anInt()
+        val fakeTextPaddingStart = forge.anInt()
+        val fakeTextPaddingEnd = forge.anInt()
+        val mockTextView: TextView = forge.aMockTextView().apply {
+            whenever(this.text).thenReturn(fakeText)
+            whenever(this.typeface).thenReturn(mock())
+            whenever(this.totalPaddingTop).thenReturn(fakeTextPaddingTop)
+            whenever(this.totalPaddingBottom).thenReturn(fakeTextPaddingBottom)
+            whenever(this.totalPaddingStart).thenReturn(fakeTextPaddingStart)
+            whenever(this.totalPaddingEnd).thenReturn(fakeTextPaddingEnd)
+            whenever(this.resources).thenReturn(mockResources)
+            whenever(this.currentTextColor).thenReturn(fakeCurrentTextColor)
+        }
+        val expectedWireframeTextPadding = MobileSegment.Padding(
+            fakeTextPaddingTop.densityNormalized(fakeMappingContext.systemInformation.screenDensity).toLong(),
+            fakeTextPaddingBottom.densityNormalized(fakeMappingContext.systemInformation.screenDensity).toLong(),
+            fakeTextPaddingStart.densityNormalized(fakeMappingContext.systemInformation.screenDensity).toLong(),
+            fakeTextPaddingEnd.densityNormalized(fakeMappingContext.systemInformation.screenDensity).toLong()
+        )
+
+        // When
+        val textWireframes = testedTextWireframeMapper.map(mockTextView, fakeMappingContext, mockAsyncJobStatusCallback)
+
+        // Then
+        val expectedWireframes = mockTextView.toTextWireframes().map {
+            it.copy(
+                text = resolveExpectedText(mockTextView),
+                textPosition = MobileSegment.TextPosition(
+                    padding = expectedWireframeTextPadding,
+                    alignment = MobileSegment.Alignment(
+                        MobileSegment.Horizontal.LEFT,
+                        MobileSegment.Vertical
+                            .CENTER
+                    )
+                )
+            )
+        }
+        assertThat(textWireframes).isEqualTo(expectedWireframes)
+    }
+
+    @Test
+    fun `M resolve a ShapeWireframe background W map {ColorDrawable background}`(
+        forge: Forge,
+        @IntForgery fakeBackgroundColor: Int,
+        @StringForgery(regex = "#[0-9A-F]{8}") fakeBackgroundColorString: String,
+        @FloatForgery(0f, 1f) fakeViewAlpha: Float
+    ) {
+        // Given
+        val mockDrawable = mock<ColorDrawable>()
+        val mockTextView = forge.aMockTextView().apply {
+            whenever(this.background).thenReturn(mockDrawable)
+            whenever(this.text).thenReturn(fakeText)
+            whenever(this.typeface).thenReturn(mock())
+            whenever(this.alpha).thenReturn(fakeViewAlpha)
+            whenever(this.resources).thenReturn(mockResources)
+            whenever(this.currentTextColor).thenReturn(fakeCurrentTextColor)
+        }
+        whenever(mockDrawableToColorMapper.mapDrawableToColor(mockDrawable)).thenReturn(fakeBackgroundColor)
+        whenever(mockColorStringFormatter.formatColorAsHexString(fakeBackgroundColor))
+            .thenReturn(fakeBackgroundColorString)
+
+        // When
+        val actualWireframes = testedTextWireframeMapper.map(
+            mockTextView,
+            fakeMappingContext,
+            mockAsyncJobStatusCallback
+        )
+
+        // Then
+        val textWireframes = mockTextView.toTextWireframes()
+
+        val backgroundWireframe = MobileSegment.Wireframe.ShapeWireframe(
+            id = fakeWireframeId,
+            x = textWireframes[0].x,
+            y = textWireframes[0].y,
+            width = mockTextView.width.densityNormalized(
+                fakeMappingContext.systemInformation.screenDensity
+            ).toLong(),
+            height = mockTextView.height.densityNormalized(
+                fakeMappingContext.systemInformation.screenDensity
+            ).toLong(),
+            shapeStyle = MobileSegment.ShapeStyle(
+                backgroundColor = fakeBackgroundColorString,
+                opacity = fakeViewAlpha,
+                cornerRadius = null
+            ),
+            border = null
+        )
+
+        val expectedWireframes = mutableListOf<MobileSegment.Wireframe>()
+        expectedWireframes.add(backgroundWireframe)
+
+        expectedWireframes.addAll(
+            mockTextView.toTextWireframes().map {
+                it.copy(
+                    text = resolveExpectedText(mockTextView)
+                )
+            }
+        )
+
+        assertThat(actualWireframes.size).isEqualTo(2)
+        assertThat(actualWireframes[0].shapeStyle()).isEqualTo(expectedWireframes[0].shapeStyle())
+        assertThat(actualWireframes[1].shapeStyle()).isNull()
+        assertThat((actualWireframes[1] as MobileSegment.Wireframe.TextWireframe).text)
+            .isEqualTo((expectedWireframes[1] as MobileSegment.Wireframe.TextWireframe).text)
+    }
+
+    @Test
+    fun `M resolve an ImageWireframe background W map {no ColorDrawable background}`(
+        forge: Forge
+    ) {
+        // Given
+        val fakeViewAlpha = forge.aFloat(min = 0f, max = 1f)
+        val mockDrawable = mock<Drawable>().also { mockDrawable ->
+            val mockConstantState = mock<Drawable.ConstantState>().also { mockState ->
+                whenever(mockState.newDrawable(anyOrNull())) doReturn mockDrawable
+            }
+            whenever(mockDrawable.constantState) doReturn mockConstantState
+        }
+        whenever(mockDrawableToColorMapper.mapDrawableToColor(mockDrawable)) doReturn null
+        val mockTextView = forge.aMockTextView().apply {
+            whenever(this.background).thenReturn(mockDrawable)
+            whenever(this.text).thenReturn(fakeText)
+            whenever(this.typeface).thenReturn(mock())
+            whenever(this.alpha).thenReturn(fakeViewAlpha)
+            whenever(this.resources).thenReturn(mockResources)
+            whenever(this.currentTextColor).thenReturn(fakeCurrentTextColor)
+        }
+        val fakeBackgroundWireframe: MobileSegment.Wireframe.ImageWireframe = forge.getForgery()
+        whenever(
+            mockImageWireframeHelper.createImageWireframe(
+                view = any(),
+                currentWireframeIndex = any(),
+                x = any(),
+                y = any(),
+                width = any(),
+                height = any(),
+                usePIIPlaceholder = any(),
+                drawable = any(),
+                shapeStyle = anyOrNull(),
+                border = anyOrNull(),
+                clipping = anyOrNull(),
+                prefix = anyOrNull(),
+                asyncJobStatusCallback = anyOrNull()
+            )
+        ).thenReturn(fakeBackgroundWireframe)
+
+        // When
+        val actualWireframes = testedTextWireframeMapper.map(
+            mockTextView,
+            fakeMappingContext,
+            mockAsyncJobStatusCallback
+        )
+
+        // Then
+        val expectedWireframes = mutableListOf<MobileSegment.Wireframe>()
+        expectedWireframes.add(fakeBackgroundWireframe)
+
+        expectedWireframes.addAll(
+            mockTextView.toTextWireframes().map {
+                it.copy(
+                    text = resolveExpectedText(mockTextView)
+                )
+            }
+        )
+
+        assertThat(actualWireframes.size).isEqualTo(2)
+        assertThat(actualWireframes[0]).isEqualTo(fakeBackgroundWireframe)
+        assertThat((actualWireframes[1] as MobileSegment.Wireframe.TextWireframe).text)
+            .isEqualTo((expectedWireframes[1] as MobileSegment.Wireframe.TextWireframe).text)
+
+        verify(mockImageWireframeHelper).createImageWireframe(
+            view = eq(mockTextView),
+            currentWireframeIndex = any(),
+            x = any(),
+            y = any(),
+            width = any(),
+            height = any(),
+            usePIIPlaceholder = any(),
+            drawable = any(),
+            asyncJobStatusCallback = eq(mockAsyncJobStatusCallback),
+            clipping = anyOrNull(),
+            shapeStyle = anyOrNull(),
+            border = anyOrNull(),
+            prefix = anyOrNull()
+        )
+    }
+
+    @Test
+    fun `M resolve a TextWireframe W map() { TextView without text, with hint }`(
+        forge: Forge,
+        @StringForgery fakeHintText: String,
+        @IntForgery(0, 0xFFFFFFF) fakeHintColor: Int,
+        @StringForgery(regex = "#[0-9A-F]{8}") fakeHintColorString: String
+    ) {
+        // Given
+        val mockColorStateList: ColorStateList = mock {
+            whenever(it.defaultColor).thenReturn(fakeHintColor)
+        }
+        val mockTextView: TextView = forge.aMockTextView().apply {
+            whenever(this.text).thenReturn("")
+            whenever(this.hint).thenReturn(fakeHintText)
+            whenever(this.hintTextColors).thenReturn(mockColorStateList)
+            whenever(this.typeface).thenReturn(mock())
+            whenever(this.resources).thenReturn(mockResources)
+            whenever(this.currentTextColor).thenReturn(fakeCurrentTextColor)
+        }
+        whenever(mockColorStringFormatter.formatColorAndAlphaAsHexString(fakeHintColor, 255))
+            .doReturn(fakeHintColorString)
+
+        // When
+        val textWireframes = testedTextWireframeMapper.map(mockTextView, fakeMappingContext, mockAsyncJobStatusCallback)
+
+        // Then
+        val expectedWireframes = mockTextView
+            .toTextWireframes()
+            .map {
+                it.copy(
+                    text = resolveExpectedText(mockTextView),
+                    textStyle = MobileSegment.TextStyle(
+                        TextViewMapper.SANS_SERIF_FAMILY_NAME,
+                        mockTextView.textSize.toLong()
+                            .densityNormalized(fakeMappingContext.systemInformation.screenDensity),
+                        fakeHintColorString
+                    )
+                )
+            }
+        assertThat(textWireframes).isEqualTo(expectedWireframes)
+    }
+
+    @Test
+    fun `M resolve a TextWireframe W map() { TextView without text, with hint, no hint color }`(
+        forge: Forge
+    ) {
+        // Given
+        val fakeHintText = forge.aString()
+        val mockTextView: TextView = forge.aMockTextView().apply {
+            whenever(this.text).thenReturn("")
+            whenever(this.hint).thenReturn(fakeHintText)
+            whenever(this.hintTextColors).thenReturn(null)
+            whenever(this.typeface).thenReturn(mock())
+            whenever(this.currentTextColor).thenReturn(fakeCurrentTextColor)
+            whenever(this.resources).thenReturn(mockResources)
+        }
+
+        // When
+        val textWireframes = testedTextWireframeMapper.map(mockTextView, fakeMappingContext, mockAsyncJobStatusCallback)
+
+        // Then
+        val expectedWireframes = mockTextView
+            .toTextWireframes()
+            .map {
+                it.copy(
+                    text = resolveExpectedText(mockTextView),
+                    textStyle = MobileSegment.TextStyle(
+                        TextViewMapper.SANS_SERIF_FAMILY_NAME,
+                        mockTextView.textSize.toLong()
+                            .densityNormalized(fakeMappingContext.systemInformation.screenDensity),
+                        fakeCurrentTextColorString
+                    )
+                )
+            }
+        assertThat(textWireframes).isEqualTo(expectedWireframes)
+    }
+
+    @Test
+    fun `M resolve an imageWireframe W map() { TextView with compoundDrawable }`(forge: Forge) {
+        // Given
+        val fakeHintText = forge.aString()
+        val fakeDrawables = arrayOf<Drawable>(
+            mock(),
+            mock(),
+            mock(),
+            mock()
+        )
+        val mockTextView: TextView = forge.aMockTextView().apply {
+            whenever(this.text).thenReturn("")
+            whenever(this.hint).thenReturn(fakeHintText)
+            whenever(this.hintTextColors).thenReturn(null)
+            whenever(this.typeface).thenReturn(mock())
+            whenever(this.currentTextColor).thenReturn(fakeCurrentTextColor)
+            whenever(this.resources).thenReturn(mockResources)
+            whenever(this.compoundDrawables).thenReturn(
+                fakeDrawables
+            )
+        }
+        val mockImageWireframe: MobileSegment.Wireframe.ImageWireframe = mock()
+        whenever(
+            mockImageWireframeHelper.createCompoundDrawableWireframes(
+                any(),
+                any(),
+                any(),
+                any()
+            )
+        ).thenReturn(mutableListOf(mockImageWireframe))
+
+        // When
+        val wireframes = testedTextWireframeMapper.map(
+            mockTextView,
+            fakeMappingContext,
+            mockAsyncJobStatusCallback
+        )
+        val imageWireframes = wireframes.filterIsInstance<MobileSegment.Wireframe.ImageWireframe>()
+
+        // Then
+        verify(mockImageWireframeHelper)
+            .createCompoundDrawableWireframes(
+                any(),
+                any(),
+                any(),
+                eq(mockAsyncJobStatusCallback)
+            )
+        assertThat(imageWireframes).isEqualTo(listOf(mockImageWireframe))
+    }
+
+    private fun resolveExpectedText(textView: TextView): String {
+        val obfuscationRule = when (fakeMappingContext.privacy) {
+            SessionReplayPrivacy.ALLOW -> AllowObfuscationRule()
+            SessionReplayPrivacy.MASK -> MaskObfuscationRule()
+            SessionReplayPrivacy.MASK_USER_INPUT -> MaskInputObfuscationRule()
+        }
+        return obfuscationRule.resolveObfuscatedValue(textView, fakeMappingContext)
     }
 }
